refactor: more ts fixes to get this to work
